+++ backend/src/routers/v1/users/index.ts
@@ -2,14 +2,7 @@ import { Request, Response } from 'express';
 import { Controller, Req, Res, Get, Put, UseBefore, Redirect } from 'routing-controllers';
 import * as passport from 'passport';
 
-import {
-  PostService,
-  UserService,
-  BlogService,
-  TistoryService,
-  NotifyService,
-  FollowService,
-} from 'src/services';
+import { PostService, UserService, BlogService, NotifyService, FollowService } from 'src/services';
 import { Authorization, ERROR, RESPONSECODE, Cursor } from 'src/utils';
 
 @Controller('/users')
@@ -107,34 +100,12 @@ export default class UsersRouter {
     return response.json({ code: RESPONSECODE.SUCCESS, data: randomUserSuggestions });
   }
 
-  @Get('/:userID/blogs')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async getUserTistory(@Req() request: Request, @Res() response: Response) {
-    const { userID } = request.params;
-    if (userID !== request.user?.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const posts = await BlogService.findUserBlogs(userID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: posts });
-  }
-
   @Get('/:userID/notifies')
   @UseBefore(passport.authenticate('jwt-registered', { session: false }))
   async getNotifyList(@Req() request: Request, @Res() response: Response) {
     const { userID } = request.params;
     const result = await NotifyService.findNotify(userID);
     return response.json({ code: RESPONSECODE.SUCCESS, data: result });
-  }
-
-  @Get('/:userID/tistory/:identity/posts/:postID')
-  @UseBefore(passport.authenticate('jwt-registered', { session: false }))
-  async getTistoryPostContent(@Req() request: Request, @Res() response: Response) {
-    const { userID, identity, postID } = request.params;
-    if (userID !== request.user?.userID) {
-      throw new Error(ERROR.PERMISSION_DENIED);
-    }
-    const postContent = await TistoryService.findPostContent(userID, identity, postID);
-    return response.json({ code: RESPONSECODE.SUCCESS, data: postContent });
   }
 
   @Put('/:userID/settings')
